memoryproject/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-environ==0.10.0
environ==1.0
psycopg2==2.9.6
sqlparse==0.4.3
tzdata==2023.3
djangorestframework
django-filter
djangorestframework-jwt
django-jazzmin
django-import-export
django-resized
django-cors-headers
django-credit-cards
django-ckeditor
django-elasticsearch-dsl-drf
django-modeltranslation
psycopg2-binary
Pillow
drf-yasg
cryptography
flake8
boto3
django-storages
Pillow
virtualenv
requests
pipreqs
PyJWT
datetime
